+++ lib/flutter_xml_widgets/lib/parsers/attr_evaluator.dart
