+++ lib/flutter_xml_widgets/lib/parsers/color_resolver.dart
