+++ lib/flutter_xml_widgets/old/utils/color_parser.dart
