+++ test/expression_evaluator_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:xml_to_dart/flutter_xml_widgets/lib/utils/expression_evaluator.dart';
+import 'package:xml_to_dart/flutter_xml_widgets/old/utils/expression_evaluator.dart';
 
 void main() {
   group('Expression Evaluation Tests', () {
